Close #6
+ Update pom.xml. Change spring-vaadin-security from 0.0.3-SNAPSHOT to 0.0.5.RELEASE
+ Change Movie POJO id type from Long to long
+ Fix all related problems with Movie id
+ Replace hashCode and equals methods in POJOs with methods from Lombok
+ Add @Transactional to all save methods in services
+ getPriceBySeanceColRow now accept long id (were int)
+ All new objects get new id after saving
+ Replace all try-catch(IncorrectResultSizeDataAccessException) with try-catch(CinemaEmptyResultException)
- remove unused listener interface from TicketSelect

+++ src/main/java/com/netcracker/cinema/dao/PriceDao.java
@@ -17,7 +17,7 @@ public interface PriceDao {
 
     void delete(Price price);
 
-    int getPriceBySeanceColRow(int seanceId, int col, int row);
+    int getPriceBySeanceColRow(long seanceId, int col, int row);
 
     Price getPriceBySeanceZone(long seanceId, long zoneId);
 }
